Rename TS file with proper convention
