+++ admin-dev/themes/new-theme/js/pages/product/combination/quantity-mode-switcher.ts
